Move \n from sl-link-postfix to sl-insert-relatedlink into drawer
The default nil sl-link-postfix causes inline links inserted with
sl-link to be followed by a newline; this is surprising to the casual
user.  It is more natural to not consider the newline part of the
postfix, and instead add the newline explicitly to end the line inside
the drawer.

Discouraging \n inside sl-link-postfix does slightly change its
semantics, but IMHO for the better.  For full flexibility and
coherence, there should probably be separate pre- and postfixes for
the inline and drawer cases.