Add capture target buffer to search targets
This one is pretty edge case but it happened to me and it was
annoying.

If calling org-super-links-link from a capture buffer that has a
target which is not included in org-agenda-files; add the target buffer
to the agenda files so that it shows in search candidates.

Without this you can't link to the buffer you called the capture from
unless it's in your agenda files.